iot-image/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv>=1.0.0
paramiko>=3.4.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
face-recognition==1.3.0
dlib==19.24.2
mediapipe>=0.10.0
aiofiles>=23.0.0
mediapipe==0.10.7
tensorflow-gpu==2.13.0
requests==2.31.0
websockets==12.0
python-dotenv==1.0.0
face_recognition
mediapipe